requirements.txt
requests==2.31.0
openai==1.14.3
python-dotenv==1.0.1
pydantic==2.6.4
requests-oauthlib==1.3.1
python-dateutil==2.9.0
typing-extensions==4.10.0 